Move document_search.rs into fs_explorer crate
