+++ crates/fs_explorer/Cargo.toml
@@ -3,6 +3,10 @@ name = "fs_explorer"
 version = "0.1.0"
 edition = "2024"
 
+[features]
+default = []
+document-conversion = ["transmutation"]
+
 [dependencies]
 anyhow = "1.0.95"
 async-trait = "0.1.85"
@@ -15,6 +19,8 @@ regex = "1.11.1"
 serde = { version = "1.0.215", features = ["derive"] }
 tracing = "0.1.40"
 path-clean = "1.0.1"
+walkdir = "2.5"
+transmutation = { version = "0.3", features = ["office"], optional = true }
 
 [dev-dependencies]
 tempfile = "3.13.0"
