+++ crates/fs_explorer/src/lib.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "document-conversion")]
+pub mod document_search;
 pub mod encoding;
 mod explorer;
 pub mod file_updater;
